+++ Iterum.PhysX/Scene.cs
@@ -92,10 +92,10 @@ internal class Scene
 
     #region Create objects
         
-    public IStaticObject CreateTerrain(Memory<float> buffer, float scale, float size, Vector3 pos, IMaterial mat = null)
+    public IStaticObject CreateTerrain(Memory<float> buffer, float scale, long size, Vector3 pos, IMaterial mat = null)
     {
         mat ??= legacySceneGlobalMaterial;
-        var obj = new TerrainObject(buffer, scale, size,  mat, pos, this);
+        var obj = new TerrainObject(buffer, scale, size, mat, pos, this);
         refs.Add(obj.Ref, obj);
 #if PHYSICS_DEBUG_LEVEL            
             Console.WriteLine($"{LogGroup} StaticObject Ref: ({obj.Ref}) created");
